Add new Reactions Query Endpoint
