+++ Sources/StreamChat/APIClient/Endpoints/MessageEndpoints.swift
@@ -50,50 +50,6 @@ extension Endpoint {
         )
     }
 
-    static func loadReactions(messageId: MessageId, pagination: Pagination) -> Endpoint<MessageReactionsPayload> {
-        .init(
-            path: .reactions(messageId),
-            method: .get,
-            queryItems: nil,
-            requiresConnectionId: false,
-            body: pagination
-        )
-    }
-
-    static func addReaction(
-        _ type: MessageReactionType,
-        score: Int,
-        enforceUnique: Bool,
-        extraData: [String: RawJSON],
-        messageId: MessageId
-    ) -> Endpoint<EmptyResponse> {
-        let body = MessageReactionRequestPayload(
-            enforceUnique: enforceUnique,
-            reaction: ReactionRequestPayload(
-                type: type,
-                score: score,
-                extraData: extraData
-            )
-        )
-        return .init(
-            path: .addReaction(messageId),
-            method: .post,
-            queryItems: nil,
-            requiresConnectionId: false,
-            body: body
-        )
-    }
-
-    static func deleteReaction(_ type: MessageReactionType, messageId: MessageId) -> Endpoint<EmptyResponse> {
-        .init(
-            path: .deleteReaction(messageId, type),
-            method: .delete,
-            queryItems: nil,
-            requiresConnectionId: false,
-            body: nil
-        )
-    }
-
     static func dispatchEphemeralMessageAction(
         cid: ChannelId,
         messageId: MessageId,
